A working rate-limiting future
This change fixes the poll interactions where the future didn't get
scheduled to be polled again.

Also adds three tests that exercise the various code paths -
delay-less allowing (the fast path), allowing right after the delay
expires, and re-scheduling a delay when the first delay expires and
the request is still not allowed.